requirements.txt
python-digitalocean==1.17.0
paramiko==2.11.0
Flask==2.1.0
gunicorn==20.1.0